+++ apps/web/src/app/partner/_components/PendingReviewPreview.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import { StatusBadge } from "@/shared/ui/StatusBadge";
 import { buttonVariants } from "@/shared/ui/Button";
 import { accidentTypeLabel } from "@/shared/model/accident-type";
-import { useReviewList } from "../review/_api/use-review-list";
+import { useReviewList } from "../_shared/api/use-review-list";
 import type { ReviewListItem } from "../_shared/model/types";
 import { SectionCard } from "./SectionCard";
 import { PendingReviewEmpty } from "./PendingReviewEmpty";
