+++ apps/web/src/app/partner/_shared/api/use-review-list.ts
@@ -4,7 +4,7 @@ import { useSuspenseQuery } from "@tanstack/react-query";
 import { reportKeys } from "@/shared/api/query-keys";
 import type { ReviewListFilter } from "@/shared/api/query-keys";
 import { GC_TIME_DEFAULT, STALE_TIME_LIST } from "@/shared/api/query-constants";
-import { getReviewList } from "../../_shared/api/get-review-list";
+import { getReviewList } from "./get-review-list";
 
 export function useReviewList(filter?: ReviewListFilter) {
   return useSuspenseQuery({
